+++ src/features/ResourceManager/components/Explorer/ListView/ListItem/index.tsx
@@ -7,7 +7,6 @@ import { isNull } from 'es-toolkit/compat';
 import { FileBoxIcon, FileText, FolderIcon } from 'lucide-react';
 import { type DragEvent, memo, useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useNavigate, useSearchParams } from 'react-router-dom';
 import { shallow } from 'zustand/shallow';
 
 import {
@@ -24,6 +23,7 @@ import { type FileListItem as FileListItemType } from '@/types/files';
 import { formatSize } from '@/utils/format';
 import { isChunkingUnsupported } from '@/utils/isChunkingUnsupported';
 
+import { useFileItemClick } from '../../hooks/useFileItemClick';
 import DropdownMenu from '../../ItemDropdown/DropdownMenu';
 import { useFileItemDropdown } from '../../ItemDropdown/useFileItemDropdown';
 import ChunksBadge from './ChunkTag';
@@ -159,9 +159,6 @@ const FileListItem = memo<FileListItemProps>(
   }) => {
     const { t } = useTranslation(['components', 'file']);
     const { message } = App.useApp();
-    const navigate = useNavigate();
-    const [, setSearchParams] = useSearchParams();
-
     // Consolidate all FileStore subscriptions with shallow equality
     const fileStoreState = useFileStore(
       (s) => ({
@@ -177,8 +174,6 @@ const FileListItem = memo<FileListItemProps>(
     const resourceManagerState = useResourceManagerStore(
       (s) => ({
         libraryId: s.libraryId,
-        setCurrentViewItemId: s.setCurrentViewItemId,
-        setMode: s.setMode,
         setPendingRenameItemId: s.setPendingRenameItemId,
       }),
       shallow,
@@ -333,40 +328,14 @@ const FileListItem = memo<FileListItemProps>(
       setRenamingValue(name);
     }, [name]);
 
-    // Memoize click handler to prevent recreation on every render
-    const handleItemClick = useCallback(() => {
-      if (isFolder) {
-        // Navigate to folder using slug-based routing (Google Drive style)
-        const folderSlug = slug || id;
-
-        if (resourceManagerState.libraryId) {
-          navigate(`/resource/library/${resourceManagerState.libraryId}/${folderSlug}`);
-        }
-      } else if (isPage) {
-        resourceManagerState.setCurrentViewItemId(id);
-        resourceManagerState.setMode('page');
-        setSearchParams(
-          (prev) => {
-            const newParams = new URLSearchParams(prev);
-            newParams.set('file', id);
-            return newParams;
-          },
-          { replace: true },
-        );
-      } else {
-        resourceManagerState.setCurrentViewItemId(id);
-        resourceManagerState.setMode('editor');
-        // Also update URL query parameter for shareable links
-        setSearchParams(
-          (prev) => {
-            const newParams = new URLSearchParams(prev);
-            newParams.set('file', id);
-            return newParams;
-          },
-          { replace: true },
-        );
-      }
-    }, [isFolder, slug, id, resourceManagerState, navigate, isPage, setSearchParams]);
+    // Use shared click handler hook
+    const handleItemClick = useFileItemClick({
+      id,
+      isFolder,
+      isPage,
+      libraryId: resourceManagerState.libraryId,
+      slug,
+    });
 
     // Auto-start renaming if this is the pending rename item
     useEffect(() => {
